move handle request support to utils
